Fixes based on @Siegrift's suggestions: - consistent file naming - remove unused/commented code - setup precommit + CI checks   - I copied Airnode's hooks and added prettier/eslint/solidity linters. - ...and other polishing stuff
