+++ .eslintignore
@@ -1,0 +1,17 @@
+**/node_modules/**
+
+# Buidler
+**/cache/**
+**/artifacts/**
+**/buidler.config.js
+
+**/.build/**
+**/.serverless/**
+**/coverage/**
+**/dist/**
+
+# Ignore the TS wrappers created by Typechain
+packages/airnode-protocol/src/contracts
+
+# Ignore JSON snippets injected into README
+packages/airnode-validator/test/fixtures
