+++ .prettierignore
@@ -1,0 +1,30 @@
+**/node_modules/**
+
+# Buidler
+**/cache/**
+**/artifacts/**
+**/buidler.config.js
+
+**/.build/**
+**/.serverless/**
+**/coverage/**
+**/dist/**
+**/.webpack/**
+
+packages/airnode-validator/test/fixtures
+packages/airnode-protocol/src/contracts
+packages/airnode-examples/deployments
+packages/airnode-examples/integration-info.json
+
+# https://github.com/protofire/solhint/issues/317
+#
+# /* solhint-disable */ comment doesn't work either because this is a parser issue
+# Prettier uses this parser for formatting
+packages/airnode-adapter/contracts/TestDecoder.sol
+
+receipt.json
+config.json
+!packages/airnode-examples/integrations/**/config.json
+
+# Third party modules
+**/.terraform/**
