+++ package.json
@@ -12,9 +12,13 @@
   },
   "description": "A stress tester for Airnode",
   "scripts": {
-    "do-test-run": "ts-node src/do-test-run.ts",
-    "logging-poc": "ts-node src/gcp_utils.ts",
-    "stack-trace-parse": "ts-node src/stack-trace-parse.ts"
+    "do-test-run": "ts-node src/do_test_run.ts",
+    "lint": "yarn run lint:eslint && yarn run lint:solhint && yarn prettier:check",
+    "lint:eslint": "eslint . --ext .js,.ts",
+    "lint:eslint:fix": "eslint . --ext .js,.ts --fix",
+    "lint:solhint": "solhint ./**/*.sol",
+    "prettier:check": "prettier --check \"./**/*.{js,sol,ts,md,yml,json}\"",
+    "prettier:write": "prettier --write \"./**/*.{js,sol,ts,md,yml,json}\" --loglevel silent"
   },
   "devDependencies": {
     "@api3/airnode-abi": "^0.2.2",
@@ -30,6 +34,9 @@
     "@aws-sdk/client-lambda": "^3.41.0",
     "@aws-sdk/client-s3": "^3.41.0",
     "@aws-sdk/types": "^3.38.0",
+    "@ethersproject/experimental": "^5.5.0",
+    "@google-cloud/functions": "^1.2.0",
+    "@google-cloud/logging": "^9.6.3",
     "@nomiclabs/hardhat-ethers": "^2.0.2",
     "@nomiclabs/hardhat-waffle": "^2.0.1",
     "@openzeppelin/contracts": "^4.3.3",
@@ -38,25 +45,30 @@
     "@types/node": "^16.11.6",
     "@types/pg": "^8.6.1",
     "@types/prompts": "^2.0.14",
+    "bottleneck": "^2.19.5",
     "chalk": "^4.1.2",
     "dotenv": "^10.0.0",
     "ethereum-waffle": "^3.4.0",
     "ethers": "^5.4.4",
+    "googleapis": "^91.0.0",
     "hardhat": "^2.7.0",
     "jest": "^27.2.5",
+    "lodash": "^4.17.21",
     "pg": "^8.7.1",
+    "prettier": "^2.5.1",
     "prompts": "^2.4.1",
     "sqlite": "^4.0.23",
     "ts-jest": "^27.0.5",
     "ts-node": "^10.2.1",
-    "typescript": "^4.3.5"
+    "typescript": "^4.3.5",
+    "@typescript-eslint/eslint-plugin": "^5.6.0",
+    "@typescript-eslint/parser": "^5.6.0",
+    "eslint": "^8.4.1",
+    "eslint-plugin-functional": "^4.0.2",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jest": "^25.3.0",
+    "pre-commit": "^1.2.2",
+    "prettier-plugin-solidity": "^1.0.0-beta.19"
   },
-  "dependencies": {
-    "@ethersproject/experimental": "^5.5.0",
-    "@google-cloud/functions": "^1.2.0",
-    "@google-cloud/logging": "^9.6.3",
-    "bottleneck": "^2.19.5",
-    "googleapis": "^91.0.0",
-    "lodash": "^4.17.21"
-  }
+  "dependencies": {}
 }
