+++ src/config_utils.ts
@@ -1,178 +1,178 @@
-import {deepCopy} from 'ethers/lib/utils';
-import {ContractsAndRequestsConfig, StressTestConfig} from './types';
-import {getStressTestConfig} from "./utils";
-import {DEFAULT_CHAIN_ID} from "./constants";
+import { deepCopy } from 'ethers/lib/utils';
+import { ContractsAndRequestsConfig } from './types';
+import { getStressTestConfig } from './utils';
+import { DEFAULT_CHAIN_ID } from './constants';
 
 const templateConfigJson = {
-    chains: [
-        {
-            authorizers: [],
-            contracts: {
-                AirnodeRrp: '${AIRNODE_RRP_ADDRESS}',
-            },
-            id: '8995',
-            providers: {
-                exampleProvider: {
-                    url: '${PROVIDER_URL}',
-                },
-            },
-            type: 'evm',
+  chains: [
+    {
+      authorizers: [],
+      contracts: {
+        AirnodeRrp: '${AIRNODE_RRP_ADDRESS}',
+      },
+      id: '8995',
+      providers: {
+        exampleProvider: {
+          url: '${PROVIDER_URL}',
         },
-    ],
-    nodeSettings: {
-        airnodeWalletMnemonic: '${AIRNODE_WALLET_MNEMONIC}',
-        heartbeat: {
-            enabled: false,
-        },
-        httpGateway: {
-            enabled: false,
-        },
-        logFormat: 'plain',
-        logLevel: 'DEBUG',
-        nodeVersion: '0.2.2',
-        stage: 'dev',
+      },
+      type: 'evm',
     },
-    triggers: {
-        rrp: [
-            {
-                endpointId: '0xf466b8feec41e9e50815e0c9dca4db1ff959637e564bb13fefa99e9f9f90453c',
-                oisTitle: 'CoinGecko basic request',
-                endpointName: 'coinMarketData',
-            },
+  ],
+  nodeSettings: {
+    airnodeWalletMnemonic: '${AIRNODE_WALLET_MNEMONIC}',
+    heartbeat: {
+      enabled: false,
+    },
+    httpGateway: {
+      enabled: false,
+    },
+    logFormat: 'plain',
+    logLevel: 'DEBUG',
+    nodeVersion: '0.2.2',
+    stage: 'dev',
+  },
+  triggers: {
+    rrp: [
+      {
+        endpointId: '0xf466b8feec41e9e50815e0c9dca4db1ff959637e564bb13fefa99e9f9f90453c',
+        oisTitle: 'CoinGecko basic request',
+        endpointName: 'coinMarketData',
+      },
+    ],
+  },
+  ois: [
+    {
+      oisFormat: '1.0.0',
+      title: 'CoinGecko basic request',
+      version: '1.0.0',
+      apiSpecifications: {
+        servers: [
+          {
+            url: 'https://mock-api.api3mock.link',
+          },
         ],
-    },
-    ois: [
-        {
-            oisFormat: '1.0.0',
-            title: 'CoinGecko basic request',
-            version: '1.0.0',
-            apiSpecifications: {
-                servers: [
-                    {
-                        url: 'https://mock-api.api3mock.link',
-                    },
-                ],
-                paths: {
-                    '/coins/{id}': {
-                        get: {
-                            parameters: [
-                                {
-                                    in: 'path',
-                                    name: 'id',
-                                },
-                                {
-                                    in: 'query',
-                                    name: 'localization',
-                                },
-                                {
-                                    in: 'query',
-                                    name: 'tickers',
-                                },
-                                {
-                                    in: 'query',
-                                    name: 'market_data',
-                                },
-                                {
-                                    in: 'query',
-                                    name: 'community_data',
-                                },
-                                {
-                                    in: 'query',
-                                    name: 'developer_data',
-                                },
-                                {
-                                    in: 'query',
-                                    name: 'sparkline',
-                                },
-                            ],
-                        },
-                    },
-                },
-                components: {
-                    securitySchemes: {},
-                },
-                security: {},
-            },
-            endpoints: [
+        paths: {
+          '/coins/{id}': {
+            get: {
+              parameters: [
                 {
-                    name: 'coinMarketData',
-                    operation: {
-                        method: 'get',
-                        path: '/coins/{id}',
-                    },
-                    fixedOperationParameters: [
-                        {
-                            operationParameter: {
-                                in: 'query',
-                                name: 'localization',
-                            },
-                            value: 'false',
-                        },
-                        {
-                            operationParameter: {
-                                in: 'query',
-                                name: 'tickers',
-                            },
-                            value: 'false',
-                        },
-                        {
-                            operationParameter: {
-                                in: 'query',
-                                name: 'market_data',
-                            },
-                            value: 'true',
-                        },
-                        {
-                            operationParameter: {
-                                in: 'query',
-                                name: 'community_data',
-                            },
-                            value: 'false',
-                        },
-                        {
-                            operationParameter: {
-                                in: 'query',
-                                name: 'developer_data',
-                            },
-                            value: 'false',
-                        },
-                        {
-                            operationParameter: {
-                                in: 'query',
-                                name: 'sparkline',
-                            },
-                            value: 'false',
-                        },
-                    ],
-                    reservedParameters: [
-                        {
-                            name: '_type',
-                            fixed: 'int256',
-                        },
-                        {
-                            name: '_path',
-                            fixed: 'market_data.current_price.usd',
-                        },
-                        {
-                            name: '_times',
-                            fixed: '1000000',
-                        },
-                    ],
-                    parameters: [
-                        {
-                            name: 'coinId',
-                            operationParameter: {
-                                in: 'path',
-                                name: 'id',
-                            },
-                        },
-                    ],
-                    testable: true,
+                  in: 'path',
+                  name: 'id',
                 },
-            ],
+                {
+                  in: 'query',
+                  name: 'localization',
+                },
+                {
+                  in: 'query',
+                  name: 'tickers',
+                },
+                {
+                  in: 'query',
+                  name: 'market_data',
+                },
+                {
+                  in: 'query',
+                  name: 'community_data',
+                },
+                {
+                  in: 'query',
+                  name: 'developer_data',
+                },
+                {
+                  in: 'query',
+                  name: 'sparkline',
+                },
+              ],
+            },
+          },
         },
-    ],
-    apiCredentials: [],
+        components: {
+          securitySchemes: {},
+        },
+        security: {},
+      },
+      endpoints: [
+        {
+          name: 'coinMarketData',
+          operation: {
+            method: 'get',
+            path: '/coins/{id}',
+          },
+          fixedOperationParameters: [
+            {
+              operationParameter: {
+                in: 'query',
+                name: 'localization',
+              },
+              value: 'false',
+            },
+            {
+              operationParameter: {
+                in: 'query',
+                name: 'tickers',
+              },
+              value: 'false',
+            },
+            {
+              operationParameter: {
+                in: 'query',
+                name: 'market_data',
+              },
+              value: 'true',
+            },
+            {
+              operationParameter: {
+                in: 'query',
+                name: 'community_data',
+              },
+              value: 'false',
+            },
+            {
+              operationParameter: {
+                in: 'query',
+                name: 'developer_data',
+              },
+              value: 'false',
+            },
+            {
+              operationParameter: {
+                in: 'query',
+                name: 'sparkline',
+              },
+              value: 'false',
+            },
+          ],
+          reservedParameters: [
+            {
+              name: '_type',
+              fixed: 'int256',
+            },
+            {
+              name: '_path',
+              fixed: 'market_data.current_price.usd',
+            },
+            {
+              name: '_times',
+              fixed: '1000000',
+            },
+          ],
+          parameters: [
+            {
+              name: 'coinId',
+              operationParameter: {
+                in: 'path',
+                name: 'id',
+              },
+            },
+          ],
+          testable: true,
+        },
+      ],
+    },
+  ],
+  apiCredentials: [],
 };
 
 /**
@@ -181,35 +181,32 @@ const templateConfigJson = {
  * @param chainRrps An array of strings representing AirnodeRrps
  */
 export const generateConfigJson = (chainRrps: ContractsAndRequestsConfig[]) => {
-    const {
-        CloudProvider,
-        ChainId
-    } = getStressTestConfig();
-    const configSource = deepCopy(templateConfigJson);
-    const config = {
-        ...configSource,
-        nodeSettings: {
-            ...configSource.nodeSettings,
-            cloudProvider: CloudProvider
-        }
+  const { CloudProvider, ChainId, NodeVersion } = getStressTestConfig();
+  const configSource = deepCopy(templateConfigJson);
+  const config = {
+    ...configSource,
+    nodeSettings: {
+      ...configSource.nodeSettings,
+      cloudProvider: CloudProvider,
+      nodeVersion: NodeVersion ? NodeVersion : 'v0.0.1',
+    },
+  };
+
+  const chains = chainRrps.map((rrp, idx) => {
+    return {
+      authorizers: [],
+      contracts: {
+        AirnodeRrp: rrp.AirnodeRrpAddress,
+      },
+      id: (ChainId ? ChainId : DEFAULT_CHAIN_ID).toString(10),
+      providers: {
+        [`provider chain ${idx}`]: {
+          url: '${PROVIDER_URL}',
+        },
+      },
+      type: 'evm',
     };
+  });
 
-    const chains = chainRrps.map((rrp, idx) => {
-        return {
-            authorizers: [],
-            contracts: {
-                AirnodeRrp: rrp.AirnodeRrpAddress,
-            },
-            id: (ChainId ? ChainId : DEFAULT_CHAIN_ID).toString(10),
-            providers: {
-                [`provider chain ${idx}`]: {
-                    url: '${PROVIDER_URL}',
-                }
-            },
-            type: 'evm',
-        };
-    });
-
-    console.log(JSON.stringify({...config, chains}, null, 2));
-    return {...config, chains};
+  return { ...config, chains };
 };
