+++ src/constants.ts
@@ -1,2 +1,1 @@
-
 export const DEFAULT_CHAIN_ID = 31337;
