+++ src/docker-poa-network.json
@@ -1,7 +1,5 @@
 {
-  "@api3/airnode-protocol/contracts/rrp/AirnodeRrp": [
-    "0x5FbDB2315678afecb367f032d93F642f64180aa3"
-  ],
+  "@api3/airnode-protocol/contracts/rrp/AirnodeRrp": ["0x5FbDB2315678afecb367f032d93F642f64180aa3"],
   "contracts/Requester": [
     "0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9",
     "0x0165878A594ca255338adfa4d48449f69242Eb8F",
