--- src/index.ts
@@ -1,9 +1,0 @@
-export * from './evm';
-export * from './utils';
-export * from './cli';
-export * from './types';
-export * from './database';
-export * from './config_utils';
-export * from './chain';
-export * from './aws_utils';
-export * from './metrics_output';
