+++ src/utils.ts
@@ -3,26 +3,27 @@ import * as os from 'os';
 import { readFileSync, writeFileSync } from 'fs';
 import { spawn } from 'child_process';
 import { randomUUID } from 'crypto';
-import { Worker } from 'worker_threads';
 import { parse as parseEnvFile } from 'dotenv';
 import { ContractsAndRequestsConfig, IntegrationInfo, StressTestConfig } from './types';
 import { generateConfigJson } from './config_utils';
 import { cliPrint } from './cli';
 import { getAirnodeWalletMnemonic } from './chain';
-import {DEFAULT_CHAIN_ID} from "./constants";
+import { DEFAULT_CHAIN_ID } from './constants';
 
 export const getGcpCredentials = () => {
   const stressTestConfig = getStressTestConfig();
-  if (stressTestConfig.CloudProvider.name === 'gcp') {
+  if (stressTestConfig.CloudProvider.type === 'gcp') {
     return [
-      ` -v ${path.join(os.homedir(), '/.config/gcloud/application_default_credentials.json:/application_default_credentials.json')} `,
-      ' -e "GOOGLE_APPLICATION_CREDENTIALS=/application_default_credentials.json" '
+      ` -v ${path.join(
+        os.homedir(),
+        '/.config/gcloud/application_default_credentials.json:/application_default_credentials.json'
+      )} `,
+      ' -e "GOOGLE_APPLICATION_CREDENTIALS=/application_default_credentials.json" ',
     ];
   }
 
   return [' ', ' '];
 };
-
 
 /**
  * @returns a provider URL based on the stress test config dictionary and optionally a request count (only applicable
@@ -32,6 +33,10 @@ export const getGcpCredentials = () => {
  * @param requestCount The number of requests the endpoint will return for getLogs, optional
  */
 export const getConfiguredProviderURL = (stressTestConfig: StressTestConfig, requestCount?: number) => {
+  if (stressTestConfig.InfuraProviderURL) {
+    return stressTestConfig.InfuraProviderURL;
+  }
+
   switch (stressTestConfig.TestType) {
     case 'MockedProvider':
       return `https://mockedrpc.api3mock.link/${requestCount ? requestCount : 100}/`;
@@ -64,7 +69,8 @@ export const contains = (a?: string, b?: string): boolean => {
  */
 export const getIntegrationInfo = (): IntegrationInfo => {
   const stressTestConfig = getStressTestConfig();
-  const integrationConfig = {
+
+  return {
     integration: 'coingecko-testable',
     airnodeType: 'local',
     network: 'docker-poa-network',
@@ -73,8 +79,6 @@ export const getIntegrationInfo = (): IntegrationInfo => {
       : 'test test test test test test test test test test test junk',
     providerUrl: getConfiguredProviderURL(stressTestConfig),
   } as IntegrationInfo;
-
-  return integrationConfig;
 };
 
 /**
@@ -90,24 +94,6 @@ export const readAirnodeSecrets = () => {
 
   return parseEnvFile(readFileSync(join(__dirname, `../integrations/${integrationInfo.integration}/secrets.env`)));
 };
-
-/**
- * @returns The contents of the "config.json" file for the current integration (throws if it doesn't exist)
- */
-export const readConfig = () => {
-  const integrationInfo = getIntegrationInfo();
-
-  const config = JSON.parse(
-    readFileSync(join(__dirname, `../integrations/${integrationInfo.integration}/config.json`)).toString()
-  );
-  return config;
-};
-
-/**
- * @param secrets The lines of the secrets file
- * @returns All the lines joined followed by a new line symbol
- */
-export const formatSecrets = (secrets: string[]) => secrets.join('\n') + '\n';
 
 /**
  * @param filename
@@ -245,9 +231,10 @@ export const pluralString = (input: number) => (input === 1 ? '' : 's');
  * @param providerOverride an optional provider override URL
  */
 export const refreshSecrets = async (requestCount?: number, providerOverride?: string) => {
-  const {ChainId} = getStressTestConfig();
-  const providerUrl = providerOverride ?
-      providerOverride : getConfiguredProviderURL(getStressTestConfig(), requestCount);
+  const { ChainId } = getStressTestConfig();
+  const providerUrl = providerOverride
+    ? providerOverride
+    : getConfiguredProviderURL(getStressTestConfig(), requestCount);
 
   //https://mitm-hardhat.api3mock.link
   const airnodeSecrets = `PROVIDER_URL=${providerUrl}
@@ -288,7 +275,7 @@ export const removeAirnode = async () => {
   const integrationPath = join(__dirname, '../integration-info.json');
   writeFileSync(integrationPath, JSON.stringify(getIntegrationInfo()));
   const secretsFilePath = join(__dirname, '../aws.env');
-  const [gcpCredsMount, gcpCredsEnv ] = getGcpCredentials();
+  const [gcpCredsMount, gcpCredsEnv] = getGcpCredentials();
 
   const deployCommand = [
     `docker run -i --rm`,
@@ -299,7 +286,7 @@ export const removeAirnode = async () => {
     `-v ${join(__dirname, '../')}:/app/output`,
     `api3/airnode-deployer:latest remove -r output/receipt.json --debug`,
   ].join(' ');
-  console.log(deployCommand);
+  cliPrint.info(deployCommand);
 
   cliPrint.info(deployCommand);
   // runShellCommand(deployCommand);
